RangeUtils: migrate logic from kbn to grafana/data (#27347)
