+++ public/app/features/variables/interval/actions.ts
@@ -1,4 +1,4 @@
-import { AppEvents } from '@grafana/data';
+import { AppEvents, rangeUtil } from '@grafana/data';
 
 import { toVariablePayload, VariableIdentifier } from '../state/types';
 import { ThunkResult } from '../../../types';
@@ -6,7 +6,6 @@ import { createIntervalOptions } from './reducer';
 import { validateVariableSelectionState } from '../state/actions';
 import { getVariable } from '../state/selectors';
 import { IntervalVariableModel } from '../types';
-import kbn from '../../../core/utils/kbn';
 import { getTimeSrv } from '../../dashboard/services/TimeSrv';
 import templateSrv from '../../templating/template_srv';
 import appEvents from '../../../core/app_events';
@@ -29,7 +28,7 @@ export const updateIntervalVariableOptions = (
 };
 
 export interface UpdateAutoValueDependencies {
-  kbn: typeof kbn;
+  calculateInterval: typeof rangeUtil.calculateInterval;
   getTimeSrv: typeof getTimeSrv;
   templateSrv: typeof templateSrv;
 }
@@ -37,14 +36,14 @@ export interface UpdateAutoValueDependencies {
 export const updateAutoValue = (
   identifier: VariableIdentifier,
   dependencies: UpdateAutoValueDependencies = {
-    kbn: kbn,
+    calculateInterval: rangeUtil.calculateInterval,
     getTimeSrv: getTimeSrv,
     templateSrv: templateSrv,
   }
 ): ThunkResult<void> => (dispatch, getState) => {
   const variableInState = getVariable<IntervalVariableModel>(identifier.id, getState());
   if (variableInState.auto) {
-    const res = dependencies.kbn.calculateInterval(
+    const res = dependencies.calculateInterval(
       dependencies.getTimeSrv().timeRange(),
       variableInState.auto_count,
       variableInState.auto_min
